litex_m2sdr_platform.py: moved to litex_m2sdr/directory to be accessible by code outside this repository
